Remove g:no_plugin_maps from config section
Having g:no_plugin_maps as a @setting means a helptag is generated for
it, and this conflicts with identical helptags created in other plugins.

Adjusted mappings file to include free text explanation of
g:no_plugin_maps.

Ensured that similar language is used to describe use of variables in
autocmds and mappings files.